refactor(ConfigPathResolver): Renames `ConfigPathDefaultResolvers` to `ConfigPathResolvers`
